Week 4 updates
Added new puzzle images, and made the game start off with a random
puzzle (I will add text file IO so the game can read and save
complete/incomplete puzzles in the next update).

Made a new class: NonogramLevel. This class will take over some of the
responsibilities from NonogramGame (reading images, creating the
solution array, determining puzzle width and height). This class was
created in preparation for the future text file saving and reading
implementation.

Added the row and column markers; they change color to show that the
player has fulfilled the requirements of the row/column.

Upon completing the puzzle, an option pane shows the colored, complete
version of the puzzle's image and a new game is started.