Week 6 updates
The game is pretty much complete at this point, so this week I made
some minor changes and added a new feature.

I added a timer that starts once the player clicks on one of the boxes.
When a level is complete, the player is now shown their completion time
and is now asked if they would like to stop or continue to the next
level.

The player's times are also now saved to the level data text file.
I also made sure to remove code I no longer needed, made sure variable
names were appropriate, and tested the code.