Week 7 Updates
This week, I made a few changes to the JavaDoc comments in all of the
classes. I also moved the timer under the help and info buttons to
improve the game's layout uniformity. Finally, I made a new text file to
hold the messages that are displayed by the help button and when there
are no levels available. Previously, the messages were held in two
variables via hard-coding, but they spanned multiple lines, and I
figured it would improve readability if there weren't two long blocks of
string text between the variables.